Commit A — behavior fixes (save report + wizard mapping)
Message (short):
fix(reporter, wizard): save individual .docx reports; apply group_display_names mapping
Message (long):
Ensure step-by-step report generation writes a .docx file to disk (handle DocxTemplate vs Document). Restore display-name enrichment by preferring group_display_names mapping from the detected design. This fixes tests expecting the Word report and user-friendly group names.
Commit B — refactor(reporter): extract metadata and ROI helpers

Message (short):
refactor(reporter): extract ROI & visualization helpers from metadata builder
Message (long):
Reduce cyclomatic complexity in Reporter by extracting ROI metric collection into _collect_roi_metrics, and splitting _append_metadata_section into _append_roi_reference_map, _append_visualizations, and _append_roi_event_log. No change in behavior; improves readability and makes next refactors safer.

+++ src/zebtrack/plugins/openvino_detector.py
@@ -350,7 +350,7 @@ def _letterbox(
         r = min(r, 1.0)
 
     ratio = r, r
-    new_unpad = int(round(shape[1] * r)), int(round(shape[0] * r))
+    new_unpad = round(shape[1] * r), round(shape[0] * r)
     dw, dh = new_shape[1] - new_unpad[0], new_shape[0] - new_unpad[1]
     if auto:
         dw, dh = np.mod(dw, stride), np.mod(dh, stride)
@@ -364,7 +364,7 @@ def _letterbox(
 
     if shape[::-1] != new_unpad:
         img = cv2.resize(img, new_unpad, interpolation=cv2.INTER_LINEAR)
-    top, bottom = int(round(dh - 0.1)), int(round(dh + 0.1))
-    left, right = int(round(dw - 0.1)), int(round(dw + 0.1))
+    top, bottom = round(dh - 0.1), round(dh + 0.1)
+    left, right = round(dw - 0.1), round(dw + 0.1)
     img = cv2.copyMakeBorder(img, top, bottom, left, right, cv2.BORDER_CONSTANT, value=color)
     return img, ratio, (dw, dh)
